Fix forcing color through termcolor
Since termcolor 2.1.0, the library will now detect whether or not the
process is running a TTY and disable outputting color if so. This broke
using the `--color` argument to force color when piping to another
process (e.g. `less`), as well as our pytest function for checking
rendered output.

The fix is ensuring that we set the environment variable FORCE_COLOR in
the cases where we want color, and might not have a TTY (e.g. in the
pytest, or when using `--color` option).